Updated README to document how to test blog generation locally
Renamed some files under `blog/`, so that these files to not get
published by Jekyll.

Moved Gemfile to the top-level directory, so that the `bundle` command
can be executed at top-level instead of inside the `blog/` directory.

Removed the `Gemfile.lock` file since it doesn't seem to be necessary for
GitHub pages to function correctly.