Added a copy of FreeNAS Mini firmware upgrade instructions
The page at [1] is now dead, so I rescued it from Google Cache [2],
saved it as HTML and added it here, for posterity.

Apparently this page is not archived by the Wayback Machine, either. So
this will be the only copy of it after Google chooses to get rid of the
cached version.

It'd be nice if FreeNAS/TrueNAS folks resurrected it, though. These
instructions are valuable, and important, for their customers like me
who visit their FreeNAS Mini's web interface infrequently, or don't pay
attention to the FreeNAS UI closely.

[1]: https://www.truenas.com/docs/hardware/legacy/mini-gen2/fn-bmc-watchdog/

[2]: https://webcache.googleusercontent.com/search?q=cache:o8wEibGpj_YJ:https://www.truenas.com/docs/hardware/legacy/mini-gen2/fn-bmc-watchdog/+&cd=1&hl=en&ct=clnk&gl=us&client=firefox-b-1-d