Updated ODT version of Resume to match PDF version
Used LibreOffice to make the edits. It took quite a while to match the
'Export as PDF' output to match the current PDF version.

The output is not an exact match, but it is very close to the PDF; just
enough for me to start editing it to add new content to it, and generate
new PDFs of the updated resume.