Redirected top-level Resume to bit.ly link
This is for counting visits to the resume, and track how many visits I'm
getting just for the URL to my resume.

This is a bit of unique gymnastics, so it needs a bit of explanation.

For a while I have used top-level file GurjeetResume.pdf to host my
resume. Recently I moved the file to the ./resume/ directory, and made
this top-level file a symlink to the new one under ./resume/.

I have lately felt the need to track page hits for just my resume, and
the easiest way seems to be to create a link in bit.ly service, and
track the hits there. But since I have been using the link [1] for a
while now, I didn't want that link to stop working. Moreover, I wanted
to keep using the same link in all my future interactions.

[1]: https://gurjeet.singh.im/GurjeetResume.pdf

The first trick I tried, but which did not work, was to replace the file
with an HTML file by the same name. The HTML had a Meta Refresh [2] tag
to redirect the visitor to the bit.ly link, which will further redirect
the visitor back to the file under ./resume/ directory. This trick did
not work because this website is currently served by Github Pages
service, and in the HTTP response it sets `Content-Type` header to the
value of `application/pdf`. This value of the header forces the
visitor's browser to render the file content as a pdf. So, because the
file is not being interpreted as HTML, the Meta Refresh tag has no
effect.

[2]: https://en.wikipedia.org/wiki/Meta_refresh

After racking my brain for some time, it occurred to me that I can use
this current trick to achieve the desired effects. I created a directory
by the name `GurjeetResume.pdf`, and in it created the file
`index.html`, which contains the Meta Refresh tag to redirect the
visitor to bit.ly. When someone visits my resume URL [1], Github Pages
serves the `index.html`, as is the default behaviour of any typical
webserver. The Meta Refresh tag in the HTML tells the browser to visit
bit.ly, which does its counting, and redirects the browser back to
`./resume/GurjeetResume.pdf`, and that finally renders as a PDF file.