Added index.html to the ./resume/ directory
I want to start sharing https://gurje.et/resume as a link to my resume.

In absence of index.html, if someone visits the above link, Github Pages
(our webserver, combined with Jekyll) serves the contents of README.md.

This index.html is a symlink to ./GurjeetResume.pdf/index.html file.

By adding index.html, and using the Meta Refresh tag in that file, we
instruct the visitor to visit our bit.ly URL (for visit counting), which
then redirects them to the actual resume file.